requirements.txt
Django==3.2.5
djangorestframework==3.12.4
psycopg2==2.9.9
drf-spectacular>=0.15.1,<0.16
requests